requirements.txt
numpy~=1.25.2
pillow~=10.0.0
gymnasium~=0.29.0
wandb~=0.15.8
tqdm~=4.65.0
scipy~=1.11.1
setuptools~=68.0.0